% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tueb_themes.R
\name{theme_tueb_1}
\alias{theme_tueb_1}
\title{Primary theme for university of tuebinge colours (for ggplot2)}
\usage{
theme_tueb_1(
  col_title = uni_tuebingen[[3]],
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{col_title}{Color of title (text) elements (optional, numeric).
Default: \code{col_title = uni_tuebingen_1[[1]]}.}

\item{base_size}{Base font size (optional, numeric). 
Default: \code{base_size = 11}.}

\item{base_family}{Base font family (optional, character). 
Default: \code{base_family = ""}. 
Options include \code{"mono"}, \code{"sans"} (default), and "serif".}

\item{base_line_size}{Base line size (optional, numeric). 
Default: \code{base_line_size = base_size/22}.}

\item{base_rect_size}{Base rectangle size (optional, numeric). 
Default: \code{base_rect_size = base_size/22}.}
}
\description{
\code{theme_tueb_1} provides a basic \bold{colorCodes} theme 
to use in \bold{ggplot2} commands.
}
\details{
The theme is lightweight and no-nonsense, but somewhat 
opinionated (e.g., in using mostly grey scales to 
allow emphasizing data points with color accents).
}
\examples{

\donttest{
  # Plotting iris dataset (using ggplot2, theme_tueb_1, and unikn colors):
  
  library('ggplot2')  # theme_tueb_1 requires ggplot2 
  
  ggplot(datasets::iris) +
    geom_jitter(aes(x = Petal.Length, y = Petal.Width, color = Species), size = 3, alpha = 2/3) +
    scale_color_manual(values = usecol(c(uni_tuebingen_1[1], uni_tuebingen_1[2], uni_tuebingen_1[3]))) +
    labs(tag = "A", title = "Iris petals",
         caption = "Data from datasets::iris") + 
    theme_tueb_1()
}

}
\seealso{
\code{\link{theme_tueb_2}} for an alternative theme.

Other plot functions: 
\code{\link{slide}()},
\code{\link{theme_grau}()},
\code{\link{theme_tueb_2}()},
\code{\link{theme_unikn}()},
\code{\link{xbox}()}
}
\concept{plot functions}
